requirements.txt
pydantic
openai
python-dotenv
openai
psycopg2-binary
chromadb